move pdf creation to child process
 - This is because instantiating multiple PDf.creates asynchronously was not
   any faster than instantiating them synchronously.
 - Forking them into a new child process appears to be faster
