+++ assets/startup/build.sh
@@ -2,6 +2,6 @@
 
 echo 'Generating PDFs'
 echo 'Installing node dependencies'
-npm i html-pdf glob jsdom js-yaml
-node _site/assets/startup/pdf-gen.js
+npm i html-pdf glob jsdom js-yaml semaphore
+node _site/assets/startup/pdfgen-parent.js
 echo 'Generating PDFs Complete'
